refactor(eval): split eval-runner so .mts is vite-node-only
