+++ scripts/eval-runner-impl.ts
@@ -1,0 +1,2579 @@
+import { createRequire } from "node:module";
+import path from "node:path";
+import { fileURLToPath, pathToFileURL } from "node:url";
+
+type EvaluatorEntry = {
+  evaluator: {
+    evalName: string;
+    projectName: string;
+  } & Record<string, unknown>;
+  reporter?: unknown;
+  paramsOverride?: Record<string, unknown>;
+};
+
+type MatrixAxis = {
+  key: string;
+  values: unknown[];
+};
+
+type EvalResult = {
+  results: Array<{ error?: unknown }>;
+  summary: unknown;
+};
+
+type ProgressReporter = {
+  start: (name: string, total: number) => void;
+  stop: (name: string) => void;
+  increment: (name: string) => void;
+  setTotal: (name: string, total: number) => void;
+};
+
+type EvalOptions = Record<string, unknown> & {
+  progress?: Partial<ProgressReporter>;
+  stream?: (data: unknown) => void;
+  onStart?: (data: unknown) => void;
+  reporter?: unknown;
+  noSendLogs?: boolean;
+};
+
+type EvalFunction = (
+  projectName: string,
+  evaluator: Record<string, unknown>,
+  options?: EvalOptions,
+) => Promise<EvalResult>;
+
+type LoginFunction = (options?: Record<string, unknown>) => Promise<unknown>;
+type InitDatasetFunction = (
+  projectOrOptions: string | Record<string, unknown>,
+  options?: Record<string, unknown>,
+) => unknown;
+type InvokeFunction = (options: Record<string, unknown>) => Promise<unknown>;
+
+type BraintrustModule = {
+  Eval?: EvalFunction;
+  login?: LoginFunction;
+  initDataset?: InitDatasetFunction;
+  invoke?: InvokeFunction;
+  _internalGetGlobalState?: () => unknown;
+  default?: BraintrustModule;
+};
+
+type GlobalEvals = {
+  functions: unknown[];
+  prompts: unknown[];
+  parameters: unknown[];
+  evaluators: Record<string, EvaluatorEntry>;
+  reporters: Record<string, unknown>;
+};
+
+type BtEvalMain = (context: BtEvalContext) => void | Promise<void>;
+
+type BtEvalContext = {
+  Eval: EvalFunction;
+  runEval: (
+    projectName: string,
+    evaluator: Record<string, unknown>,
+    options?: EvalOptions,
+  ) => Promise<EvalResult>;
+  runRegisteredEvals: () => Promise<boolean>;
+  makeEvalOptions: (
+    evaluatorName: string,
+    options?: EvalOptions,
+  ) => EvalOptions | undefined;
+  sendConsole: (message: string, stream?: "stdout" | "stderr") => void;
+  sendEvent: (event: string, data: unknown) => void;
+};
+
+type SseWriter = {
+  send: (event: string, data: unknown) => void;
+  close: () => void;
+};
+
+type EvalFilter = {
+  path: string[];
+  pattern: RegExp;
+};
+
+type SerializedEvalFilter = {
+  path: string[];
+  pattern: string;
+};
+
+type EvalScoreSpec = {
+  name: string;
+  function_id: Record<string, unknown>;
+};
+
+type EvalRequest = {
+  name: string;
+  parameters?: Record<string, unknown>;
+  parent?: unknown;
+  experiment_name?: string;
+  project_id?: string;
+  data:
+    | { data: unknown[] }
+    | { project_name: string; dataset_name: string; _internal_btql?: unknown }
+    | { project_id: string; dataset_name: string; _internal_btql?: unknown };
+  scores?: EvalScoreSpec[];
+};
+
+type RunnerConfig = {
+  jsonl: boolean;
+  list: boolean;
+  terminateOnFailure: boolean;
+  filters: EvalFilter[];
+  first: number | null;
+  sample: number | null;
+  sampleSeed: number | null;
+  devMode: "list" | "eval" | null;
+  devRequestJson: string | null;
+  params: Record<string, unknown> | null;
+  matrix: MatrixAxis[] | null;
+};
+
+type EvalRunner = {
+  Eval: EvalFunction;
+  sse: SseWriter | null;
+  login?: LoginFunction;
+  initDataset?: InitDatasetFunction;
+  invoke?: InvokeFunction;
+  runEval: (
+    projectName: string,
+    evaluator: Record<string, unknown>,
+    options?: EvalOptions,
+    paramsOverride?: Record<string, unknown>,
+  ) => Promise<EvalResult>;
+  runRegisteredEvals: (evaluators: EvaluatorEntry[]) => Promise<boolean>;
+  makeEvalOptions: (
+    evaluatorName: string,
+    options?: EvalOptions,
+  ) => EvalOptions | undefined;
+  finish: (ok: boolean) => void;
+  noSendLogs: boolean;
+  parseParent: ParseParentFunction | null;
+  getState: (() => unknown) | null;
+};
+
+type ParameterContainerSerializer = (parameters: unknown) => unknown;
+type PromptDefinitionSerializer = (prompt: unknown) => unknown;
+type ZodSchemaSerializer = (schema: unknown) => Record<string, unknown>;
+type ParseParentFunction = (parent: unknown) => string | undefined;
+
+type ParameterSerializationHelpers = {
+  sdkSerializeParameters: ParameterContainerSerializer | null;
+  promptDefinitionToPromptData: PromptDefinitionSerializer | null;
+  zodToJsonSchema: ZodSchemaSerializer | null;
+};
+
+declare global {
+  // eslint-disable-next-line no-var
+  var _evals: GlobalEvals | undefined;
+  // eslint-disable-next-line no-var
+  var _lazy_load: boolean | undefined;
+  // eslint-disable-next-line no-var
+  var __inherited_braintrust_state: unknown;
+}
+
+function isObject(value: unknown): value is Record<string, unknown> {
+  return typeof value === "object" && value !== null;
+}
+
+function isBraintrustModule(value: unknown): value is BraintrustModule {
+  return isObject(value) && ("Eval" in value || "login" in value);
+}
+
+function normalizeBraintrustModule(value: unknown): BraintrustModule {
+  if (isBraintrustModule(value)) {
+    return value;
+  }
+  if (isObject(value) && isBraintrustModule(value.default)) {
+    return value.default;
+  }
+  throw new Error("Unable to load braintrust module.");
+}
+
+function normalizeFiles(files: string[]): string[] {
+  return files.map((file) => path.resolve(process.cwd(), file));
+}
+
+function envFlag(name: string): boolean {
+  const value = process.env[name];
+  if (!value) {
+    return false;
+  }
+  const normalized = value.toLowerCase();
+  return !["0", "false", "no", "off", ""].includes(normalized);
+}
+
+function serializeJSONWithPlainString(value: unknown): string {
+  if (typeof value === "string") {
+    return value;
+  }
+  return JSON.stringify(value);
+}
+
+function parseSerializedFilters(serialized: string | undefined): EvalFilter[] {
+  if (!serialized) {
+    return [];
+  }
+
+  try {
+    const parsed = JSON.parse(serialized);
+    if (!Array.isArray(parsed)) {
+      throw new Error("BT_EVAL_FILTER_PARSED must be a JSON array.");
+    }
+    return parsed.map((value) => {
+      if (!isObject(value)) {
+        throw new Error(
+          "BT_EVAL_FILTER_PARSED entries must be objects with {path, pattern}.",
+        );
+      }
+      const { path: rawPath, pattern: rawPattern } =
+        value as SerializedEvalFilter;
+      if (
+        !Array.isArray(rawPath) ||
+        !rawPath.every((part) => typeof part === "string")
+      ) {
+        throw new Error(
+          "BT_EVAL_FILTER_PARSED entry path must be an array of strings.",
+        );
+      }
+      if (typeof rawPattern !== "string") {
+        throw new Error(
+          "BT_EVAL_FILTER_PARSED entry pattern must be a string.",
+        );
+      }
+      return {
+        path: rawPath,
+        pattern: new RegExp(rawPattern),
+      };
+    });
+  } catch (err) {
+    throw new Error(
+      `Invalid BT_EVAL_FILTER_PARSED value: ${err instanceof Error ? err.message : String(err)}`,
+    );
+  }
+}
+
+function parseDevMode(value: string | undefined): "list" | "eval" | null {
+  if (!value) {
+    return null;
+  }
+  if (value === "list" || value === "eval") {
+    return value;
+  }
+  throw new Error(`Invalid BT_EVAL_DEV_MODE value: ${value}`);
+}
+
+function parsePositiveIntegerEnv(name: string): number | null {
+  const value = process.env[name];
+  if (!value) {
+    return null;
+  }
+  if (!/^[0-9]+$/.test(value)) {
+    throw new Error(`${name} must be a positive integer.`);
+  }
+  const parsed = Number(value);
+  if (!Number.isSafeInteger(parsed) || parsed < 1) {
+    throw new Error(`${name} must be a positive integer.`);
+  }
+  return parsed;
+}
+
+function parseIntegerEnv(name: string): number | null {
+  const value = process.env[name];
+  if (!value) {
+    return null;
+  }
+  if (!/^-?[0-9]+$/.test(value)) {
+    throw new Error(`${name} must be an integer.`);
+  }
+  const parsed = Number(value);
+  if (!Number.isSafeInteger(parsed)) {
+    throw new Error(`${name} must be an integer.`);
+  }
+  return parsed;
+}
+
+function parseParamsJson(
+  raw: string | undefined,
+): Record<string, unknown> | null {
+  if (!raw) {
+    return null;
+  }
+  let parsed: unknown;
+  try {
+    parsed = JSON.parse(raw);
+  } catch {
+    throw new Error(`BT_EVAL_PARAMS_JSON is not valid JSON: ${raw}`);
+  }
+  if (!parsed || typeof parsed !== "object" || Array.isArray(parsed)) {
+    throw new Error("BT_EVAL_PARAMS_JSON must be a JSON object.");
+  }
+  return parsed as Record<string, unknown>;
+}
+
+function getDeclaredParameterKeys(parameters: unknown): Set<string> | null {
+  if (!isObject(parameters)) {
+    return null;
+  }
+  if (Reflect.get(parameters, "__braintrust_parameters_marker") === true) {
+    const schema = Reflect.get(parameters, "schema");
+    if (isObject(schema)) {
+      const properties = Reflect.get(schema, "properties");
+      if (isObject(properties)) {
+        return new Set(Object.keys(properties));
+      }
+    }
+    return null;
+  }
+  if (isObject(Reflect.get(parameters, "schema"))) {
+    const schema = Reflect.get(parameters, "schema") as Record<string, unknown>;
+    const properties = Reflect.get(schema, "properties");
+    if (isObject(properties)) {
+      return new Set(Object.keys(properties));
+    }
+  }
+  return new Set(Object.keys(parameters));
+}
+
+function filterParamsForEvaluator(
+  params: Record<string, unknown>,
+  evaluatorParameters: unknown,
+): Record<string, unknown> {
+  const declared = getDeclaredParameterKeys(evaluatorParameters);
+  if (declared === null) {
+    return params;
+  }
+  const filtered: Record<string, unknown> = {};
+  for (const key of Object.keys(params)) {
+    if (declared.has(key)) {
+      filtered[key] = params[key];
+    }
+  }
+  return filtered;
+}
+
+function parseMatrixJson(raw: string | undefined): MatrixAxis[] | null {
+  if (!raw) {
+    return null;
+  }
+  let parsed: unknown;
+  try {
+    parsed = JSON.parse(raw);
+  } catch {
+    throw new Error(`BT_EVAL_MATRIX_JSON is not valid JSON: ${raw}`);
+  }
+  if (!Array.isArray(parsed)) {
+    throw new Error(
+      "BT_EVAL_MATRIX_JSON must be a JSON array of {key, values}.",
+    );
+  }
+  const axes: MatrixAxis[] = [];
+  for (const entry of parsed) {
+    if (!isObject(entry)) {
+      throw new Error(
+        "BT_EVAL_MATRIX_JSON entries must be objects with key and values.",
+      );
+    }
+    const key = Reflect.get(entry, "key");
+    const values = Reflect.get(entry, "values");
+    if (typeof key !== "string" || key.length === 0) {
+      throw new Error("BT_EVAL_MATRIX_JSON entry missing string key.");
+    }
+    if (!Array.isArray(values) || values.length === 0) {
+      throw new Error(
+        `BT_EVAL_MATRIX_JSON axis "${key}" must have a non-empty values array.`,
+      );
+    }
+    axes.push({ key, values });
+  }
+  return axes.length > 0 ? axes : null;
+}
+
+function matrixCombinations(
+  axes: MatrixAxis[],
+): Array<Array<[string, unknown]>> {
+  let combos: Array<Array<[string, unknown]>> = [[]];
+  for (const axis of axes) {
+    const next: Array<Array<[string, unknown]>> = [];
+    for (const existing of combos) {
+      for (const value of axis.values) {
+        next.push([...existing, [axis.key, value]]);
+      }
+    }
+    combos = next;
+  }
+  return combos;
+}
+
+function formatMatrixLabel(combo: Array<[string, unknown]>): string {
+  return combo
+    .map(([key, value]) => `${key}=${JSON.stringify(value)}`)
+    .join(", ");
+}
+
+function readRunnerConfig(): RunnerConfig {
+  return {
+    jsonl: envFlag("BT_EVAL_JSONL"),
+    list: envFlag("BT_EVAL_LIST"),
+    terminateOnFailure: envFlag("BT_EVAL_TERMINATE_ON_FAILURE"),
+    filters: parseSerializedFilters(process.env.BT_EVAL_FILTER_PARSED),
+    first: parsePositiveIntegerEnv("BT_EVAL_FIRST"),
+    sample: parsePositiveIntegerEnv("BT_EVAL_SAMPLE"),
+    sampleSeed: parseIntegerEnv("BT_EVAL_SAMPLE_SEED"),
+    devMode: parseDevMode(process.env.BT_EVAL_DEV_MODE),
+    devRequestJson: process.env.BT_EVAL_DEV_REQUEST_JSON ?? null,
+    params: parseParamsJson(process.env.BT_EVAL_PARAMS_JSON),
+    matrix: parseMatrixJson(process.env.BT_EVAL_MATRIX_JSON),
+  };
+}
+
+const runtimeRequire = createRequire(
+  process.argv[1] ?? path.join(process.cwd(), "package.json"),
+);
+const fsMutable = runtimeRequire("node:fs") as typeof import("node:fs");
+const moduleMutable = (() => {
+  try {
+    return runtimeRequire("node:module") as Record<string, unknown>;
+  } catch {
+    return {};
+  }
+})();
+
+type NetModule = {
+  createConnection: (options: Record<string, unknown>) => {
+    writable: boolean;
+    end: () => void;
+    setNoDelay: (value?: boolean) => void;
+    on: (event: string, listener: (...args: unknown[]) => void) => void;
+    write: (data: string) => void;
+  };
+};
+
+const dependencyFiles = new Set<string>();
+const DEPENDENCY_EXTENSIONS = new Set([
+  ".ts",
+  ".tsx",
+  ".js",
+  ".jsx",
+  ".mjs",
+  ".cjs",
+  ".mts",
+  ".cts",
+  ".json",
+]);
+const IGNORED_DEPENDENCY_SEGMENTS = [
+  "/node_modules/",
+  "/.git/",
+  "/.venv/",
+  "/__pycache__/",
+  "/site-packages/",
+  "/dist-packages/",
+];
+const STATIC_IMPORT_PATTERN =
+  /(?:import|export)\s+(?:[^"'`]*?\sfrom\s*)?["'`]([^"'`]+)["'`]|import\s*\(\s*["'`]([^"'`]+)["'`]\s*\)|require\s*\(\s*["'`]([^"'`]+)["'`]\s*\)/g;
+
+function toDependencyPath(input: unknown): string | null {
+  try {
+    if (input instanceof URL) {
+      return fileURLToPath(input);
+    }
+    if (Buffer.isBuffer(input)) {
+      return path.resolve(process.cwd(), input.toString());
+    }
+    if (typeof input !== "string") {
+      return null;
+    }
+    if (input.startsWith("file://")) {
+      return fileURLToPath(input);
+    }
+    return path.isAbsolute(input)
+      ? path.normalize(input)
+      : path.resolve(process.cwd(), input);
+  } catch {
+    return null;
+  }
+}
+
+function shouldIgnoreDependencyPath(filePath: string): boolean {
+  const normalized = filePath.replaceAll("\\", "/");
+  return IGNORED_DEPENDENCY_SEGMENTS.some((segment) =>
+    normalized.includes(segment),
+  );
+}
+
+function maybeRecordDependency(input: unknown) {
+  const filePath = toDependencyPath(input);
+  if (!filePath || shouldIgnoreDependencyPath(filePath)) {
+    return;
+  }
+
+  const extension = path.extname(filePath).toLowerCase();
+  if (!DEPENDENCY_EXTENSIONS.has(extension)) {
+    return;
+  }
+
+  try {
+    if (fsMutable.statSync(filePath).isFile()) {
+      dependencyFiles.add(filePath);
+    }
+  } catch {
+    // Ignore inaccessible or non-file inputs.
+  }
+}
+
+function maybeRecordDependencyFromSpecifier(
+  specifier: string,
+  resolveDir?: string,
+) {
+  if (
+    specifier.startsWith("node:") ||
+    specifier.startsWith("bun:") ||
+    specifier.startsWith("npm:")
+  ) {
+    return;
+  }
+
+  if (
+    specifier.startsWith("./") ||
+    specifier.startsWith("../") ||
+    specifier.startsWith("/") ||
+    specifier.startsWith("file://")
+  ) {
+    const baseDir = resolveDir ?? process.cwd();
+    const candidate = specifier.startsWith("file://")
+      ? specifier
+      : path.resolve(baseDir, specifier);
+    maybeRecordDependency(candidate);
+  }
+}
+
+function collectStaticLocalDependencies(entryFiles: string[]) {
+  const queue = [...entryFiles];
+  const visited = new Set<string>();
+
+  while (queue.length > 0) {
+    const file = queue.pop();
+    if (!file) {
+      continue;
+    }
+    const absolute = path.resolve(file);
+    if (visited.has(absolute)) {
+      continue;
+    }
+    visited.add(absolute);
+    maybeRecordDependency(absolute);
+
+    let source = "";
+    try {
+      source = fsMutable.readFileSync(absolute, "utf8");
+    } catch {
+      continue;
+    }
+
+    STATIC_IMPORT_PATTERN.lastIndex = 0;
+    let match: RegExpExecArray | null;
+    while ((match = STATIC_IMPORT_PATTERN.exec(source)) !== null) {
+      const specifier = match[1] ?? match[2] ?? match[3];
+      if (!specifier) {
+        continue;
+      }
+      const resolved = resolveLocalSpecifier(absolute, specifier);
+      if (!resolved) {
+        continue;
+      }
+      maybeRecordDependency(resolved);
+      if (!visited.has(resolved)) {
+        queue.push(resolved);
+      }
+    }
+  }
+}
+
+function resolveLocalSpecifier(
+  fromFile: string,
+  specifier: string,
+): string | null {
+  if (
+    !specifier.startsWith("./") &&
+    !specifier.startsWith("../") &&
+    !specifier.startsWith("/") &&
+    !specifier.startsWith("file://")
+  ) {
+    return null;
+  }
+
+  const fromDir = path.dirname(fromFile);
+  const base = specifier.startsWith("file://")
+    ? fileURLToPath(specifier)
+    : specifier.startsWith("/")
+      ? path.normalize(specifier)
+      : path.resolve(fromDir, specifier);
+
+  const candidates = [base];
+  if (!path.extname(base)) {
+    for (const ext of DEPENDENCY_EXTENSIONS) {
+      candidates.push(`${base}${ext}`);
+    }
+    for (const ext of DEPENDENCY_EXTENSIONS) {
+      candidates.push(path.join(base, `index${ext}`));
+    }
+  }
+
+  for (const candidate of candidates) {
+    try {
+      if (fsMutable.statSync(candidate).isFile()) {
+        return path.normalize(candidate);
+      }
+    } catch {
+      continue;
+    }
+  }
+
+  return null;
+}
+
+function installNodeModuleHooks() {
+  const registerHooks = moduleMutable.registerHooks as
+    | ((hooks: Record<string, (...args: unknown[]) => unknown>) => void)
+    | undefined;
+  if (typeof registerHooks !== "function") {
+    return;
+  }
+
+  registerHooks({
+    resolve: (specifier, context, nextResolve) => {
+      const next = nextResolve as (
+        specifier: unknown,
+        context: Record<string, unknown>,
+      ) => { url?: string } & Record<string, unknown>;
+      const ctx = (context ?? {}) as Record<string, unknown>;
+      const result = next(specifier, ctx);
+      const resolvedUrl = result?.url;
+      if (typeof resolvedUrl === "string") {
+        maybeRecordDependency(resolvedUrl);
+      } else if (typeof specifier === "string") {
+        const resolveDir =
+          typeof ctx.parentURL === "string" &&
+          ctx.parentURL.startsWith("file://")
+            ? path.dirname(fileURLToPath(ctx.parentURL))
+            : undefined;
+        maybeRecordDependencyFromSpecifier(specifier, resolveDir);
+      }
+      return result;
+    },
+  });
+}
+
+function installBunModuleHooks() {
+  const bun = (globalThis as { Bun?: Record<string, unknown> }).Bun as
+    | {
+        plugin?: (plugin: {
+          name: string;
+          setup: (build: Record<string, unknown>) => void;
+        }) => void;
+      }
+    | undefined;
+  if (!bun || typeof bun.plugin !== "function") {
+    return;
+  }
+
+  bun.plugin({
+    name: "bt-eval-dependency-tracker",
+    setup: (build: Record<string, unknown>) => {
+      const onResolve = build.onResolve as
+        | ((
+            options: { filter: RegExp },
+            callback: (args: Record<string, unknown>) => unknown,
+          ) => void)
+        | undefined;
+      if (typeof onResolve === "function") {
+        onResolve({ filter: /.*/ }, (args) => {
+          const specifier = args.path;
+          const resolveDir =
+            typeof args.resolveDir === "string"
+              ? args.resolveDir
+              : process.cwd();
+          if (typeof specifier === "string") {
+            maybeRecordDependencyFromSpecifier(specifier, resolveDir);
+          }
+          return null;
+        });
+      }
+    },
+  });
+}
+
+function installDependencyTracking() {
+  installNodeModuleHooks();
+  installBunModuleHooks();
+
+  const fsPatched = fsMutable as unknown as Record<string, unknown>;
+  const originalReadFileSync = fsMutable.readFileSync.bind(fsMutable);
+  Reflect.set(
+    fsPatched,
+    "readFileSync",
+    (file: unknown, ...args: unknown[]) => {
+      maybeRecordDependency(file);
+      const callArgs = [file, ...args] as unknown[];
+      return Reflect.apply(
+        originalReadFileSync as (...params: unknown[]) => unknown,
+        fsMutable,
+        callArgs,
+      );
+    },
+  );
+
+  const originalReadFile = fsMutable.readFile.bind(fsMutable);
+  Reflect.set(fsPatched, "readFile", (file: unknown, ...args: unknown[]) => {
+    maybeRecordDependency(file);
+    const callArgs = [file, ...args] as unknown[];
+    return Reflect.apply(
+      originalReadFile as (...params: unknown[]) => unknown,
+      fsMutable,
+      callArgs,
+    );
+  });
+
+  const originalPromisesReadFile = fsMutable.promises.readFile.bind(
+    fsMutable.promises,
+  );
+  const fsPromisesPatched = fsMutable.promises as unknown as Record<
+    string,
+    unknown
+  >;
+  Reflect.set(
+    fsPromisesPatched,
+    "readFile",
+    async (file: unknown, ...args: unknown[]) => {
+      maybeRecordDependency(file);
+      const callArgs = [file, ...args] as unknown[];
+      return Reflect.apply(
+        originalPromisesReadFile as (...params: unknown[]) => Promise<unknown>,
+        fsMutable.promises,
+        callArgs,
+      );
+    },
+  );
+}
+
+function collectRequireCacheDependencies() {
+  const cache = runtimeRequire.cache as Record<
+    string,
+    { filename?: string } | undefined
+  >;
+  if (!cache) {
+    return;
+  }
+  for (const [cacheKey, moduleValue] of Object.entries(cache)) {
+    maybeRecordDependency(moduleValue?.filename ?? cacheKey);
+  }
+}
+
+async function collectDenoInfoDependencies(files: string[]) {
+  const deno = (globalThis as Record<string, unknown>).Deno as
+    | {
+        Command?: new (
+          command: string,
+          options: Record<string, unknown>,
+        ) => {
+          output: () => Promise<{
+            success: boolean;
+            stdout: Uint8Array;
+          }>;
+        };
+      }
+    | undefined;
+  if (!deno || typeof deno.Command !== "function") {
+    return;
+  }
+
+  for (const file of files) {
+    try {
+      const cmd = new deno.Command("deno", {
+        args: ["info", "--json", file],
+        stdout: "piped",
+        stderr: "null",
+      });
+      const output = await cmd.output();
+      if (!output.success) {
+        continue;
+      }
+      const parsed = JSON.parse(new TextDecoder().decode(output.stdout));
+      collectFileUrlsFromJson(parsed);
+    } catch {
+      continue;
+    }
+  }
+}
+
+function collectFileUrlsFromJson(value: unknown) {
+  if (typeof value === "string") {
+    maybeRecordDependency(value);
+    return;
+  }
+  if (Array.isArray(value)) {
+    for (const item of value) {
+      collectFileUrlsFromJson(item);
+    }
+    return;
+  }
+  if (!value || typeof value !== "object") {
+    return;
+  }
+  for (const child of Object.values(value)) {
+    collectFileUrlsFromJson(child);
+  }
+}
+
+function collectDependencyFiles(): string[] {
+  return Array.from(dependencyFiles).sort();
+}
+
+function serializeSseEvent(event: { event?: string; data: string }): string {
+  return (
+    Object.entries(event)
+      .filter(([_key, value]) => value !== undefined)
+      .map(([key, value]) => `${key}: ${value}`)
+      .join("\n") + "\n\n"
+  );
+}
+
+function createSseWriter(): SseWriter | null {
+  const netModule = (() => {
+    try {
+      return runtimeRequire("node:net") as NetModule;
+    } catch {
+      return null;
+    }
+  })();
+
+  const sock = process.env.BT_EVAL_SSE_SOCK;
+  if (sock) {
+    if (!netModule) {
+      return null;
+    }
+    let socket: ReturnType<NetModule["createConnection"]>;
+    try {
+      socket = netModule.createConnection({ path: sock });
+    } catch (err) {
+      console.error(
+        `Failed to connect to SSE socket: ${
+          err instanceof Error ? err.message : String(err)
+        }`,
+      );
+      return null;
+    }
+    socket.on("error", (err) => {
+      const message = err instanceof Error ? err.message : String(err);
+      console.error(`Failed to connect to SSE socket: ${message}`);
+    });
+    const send = (event: string, payload: unknown) => {
+      if (!socket.writable) {
+        return;
+      }
+      const data =
+        typeof payload === "string" ? payload : JSON.stringify(payload);
+      socket.write(serializeSseEvent({ event, data }));
+    };
+    const close = () => {
+      socket.end();
+    };
+    return { send, close };
+  }
+
+  const addr = process.env.BT_EVAL_SSE_ADDR;
+  if (!addr) {
+    return null;
+  }
+
+  const [host, portStr] = addr.split(":");
+  const port = Number(portStr);
+  if (!host || !Number.isFinite(port)) {
+    throw new Error(`Invalid BT_EVAL_SSE_ADDR: ${addr}`);
+  }
+
+  if (!netModule) {
+    return null;
+  }
+
+  let socket: ReturnType<NetModule["createConnection"]>;
+  try {
+    socket = netModule.createConnection({ host, port });
+  } catch (err) {
+    console.error(
+      `Failed to connect to SSE address ${addr}: ${
+        err instanceof Error ? err.message : String(err)
+      }`,
+    );
+    return null;
+  }
+  socket.setNoDelay(true);
+  socket.on("error", (err) => {
+    const message = err instanceof Error ? err.message : String(err);
+    console.error(`Failed to connect to SSE address ${addr}: ${message}`);
+  });
+
+  const send = (event: string, payload: unknown) => {
+    if (!socket.writable) {
+      return;
+    }
+    const data =
+      typeof payload === "string" ? payload : JSON.stringify(payload);
+    socket.write(serializeSseEvent({ event, data }));
+  };
+
+  const close = () => {
+    socket.end();
+  };
+
+  return { send, close };
+}
+
+function initRegistry() {
+  globalThis._evals = {
+    functions: [],
+    prompts: [],
+    parameters: [],
+    evaluators: {},
+    reporters: {},
+  };
+  globalThis._lazy_load = true;
+}
+
+function ensureBraintrustAvailable() {
+  resolveBraintrustPath();
+}
+
+function resolveBraintrustPath(): string {
+  const files = normalizeFiles(process.argv.slice(2));
+  for (const file of files) {
+    try {
+      const require = createRequire(pathToFileURL(file).href);
+      return require.resolve("braintrust");
+    } catch {
+      continue;
+    }
+  }
+
+  try {
+    const require = createRequire(process.cwd() + "/");
+    return require.resolve("braintrust");
+  } catch {
+    const message =
+      "Unable to resolve the `braintrust` package. " +
+      "Please install it in your project (e.g. `pnpm add braintrust` or `npm install braintrust`).";
+    throw new Error(message);
+  }
+}
+
+async function loadBraintrust() {
+  const cjsPath = resolveBraintrustPath();
+  const cjsUrl = pathToFileURL(cjsPath).href;
+
+  try {
+    const mod: unknown = await import(cjsUrl);
+    return normalizeBraintrustModule(mod);
+  } catch {}
+
+  const esmPath = cjsPath.replace(/\.js$/, ".mjs");
+  if (esmPath !== cjsPath && fsMutable.existsSync(esmPath)) {
+    try {
+      const mod: unknown = await import(pathToFileURL(esmPath).href);
+      return normalizeBraintrustModule(mod);
+    } catch {}
+  }
+
+  const require = createRequire(cjsUrl);
+  const mod: unknown = require(cjsPath);
+  return normalizeBraintrustModule(mod);
+}
+
+function extractParameterSerializer(
+  mod: unknown,
+): ParameterContainerSerializer | null {
+  if (!isObject(mod)) {
+    return null;
+  }
+  const candidate = Reflect.get(mod, "serializeRemoteEvalParametersContainer");
+  if (typeof candidate === "function") {
+    return candidate as ParameterContainerSerializer;
+  }
+  const defaultExport = Reflect.get(mod, "default");
+  if (isObject(defaultExport)) {
+    const fromDefault = Reflect.get(
+      defaultExport,
+      "serializeRemoteEvalParametersContainer",
+    );
+    if (typeof fromDefault === "function") {
+      return fromDefault as ParameterContainerSerializer;
+    }
+  }
+  return null;
+}
+
+function extractPromptDefinitionSerializer(
+  mod: unknown,
+): PromptDefinitionSerializer | null {
+  if (!isObject(mod)) {
+    return null;
+  }
+  const candidate = Reflect.get(mod, "promptDefinitionToPromptData");
+  if (typeof candidate === "function") {
+    return candidate as PromptDefinitionSerializer;
+  }
+  const defaultExport = Reflect.get(mod, "default");
+  if (isObject(defaultExport)) {
+    const fromDefault = Reflect.get(
+      defaultExport,
+      "promptDefinitionToPromptData",
+    );
+    if (typeof fromDefault === "function") {
+      return fromDefault as PromptDefinitionSerializer;
+    }
+  }
+  return null;
+}
+
+function isZodV4Schema(schema: unknown): boolean {
+  return (
+    isObject(schema) &&
+    "_zod" in schema &&
+    Reflect.get(schema, "_zod") !== undefined
+  );
+}
+
+function normalizeJsonSchema(value: unknown): Record<string, unknown> {
+  if (isObject(value)) {
+    return value;
+  }
+  return {};
+}
+
+function loadZodSchemaSerializer(
+  braintrustResolvedPath: string,
+): ZodSchemaSerializer | null {
+  const requireFromBraintrust = createRequire(
+    pathToFileURL(braintrustResolvedPath).href,
+  );
+
+  let zodToJsonSchemaV3: ((schema: unknown) => unknown) | null = null;
+  try {
+    const zodToJsonSchemaModule: unknown =
+      requireFromBraintrust("zod-to-json-schema");
+    if (typeof zodToJsonSchemaModule === "function") {
+      zodToJsonSchemaV3 = zodToJsonSchemaModule as (schema: unknown) => unknown;
+    } else if (isObject(zodToJsonSchemaModule)) {
+      const direct = Reflect.get(zodToJsonSchemaModule, "zodToJsonSchema");
+      if (typeof direct === "function") {
+        zodToJsonSchemaV3 = direct as (schema: unknown) => unknown;
+      } else {
+        const nestedDefault = Reflect.get(zodToJsonSchemaModule, "default");
+        if (typeof nestedDefault === "function") {
+          zodToJsonSchemaV3 = nestedDefault as (schema: unknown) => unknown;
+        } else if (isObject(nestedDefault)) {
+          const fromDefault = Reflect.get(nestedDefault, "zodToJsonSchema");
+          if (typeof fromDefault === "function") {
+            zodToJsonSchemaV3 = fromDefault as (schema: unknown) => unknown;
+          }
+        }
+      }
+    }
+  } catch {
+    zodToJsonSchemaV3 = null;
+  }
+
+  let zodToJsonSchemaV4:
+    | ((schema: unknown, options?: { target?: string }) => unknown)
+    | null = null;
+  try {
+    const zodV4Module: unknown = requireFromBraintrust("zod/v4");
+    if (isObject(zodV4Module)) {
+      const direct = Reflect.get(zodV4Module, "toJSONSchema");
+      if (typeof direct === "function") {
+        zodToJsonSchemaV4 = direct as (
+          schema: unknown,
+          options?: { target?: string },
+        ) => unknown;
+      }
+    }
+  } catch {
+    zodToJsonSchemaV4 = null;
+  }
+
+  if (!zodToJsonSchemaV3 && !zodToJsonSchemaV4) {
+    return null;
+  }
+
+  return (schema: unknown): Record<string, unknown> => {
+    try {
+      if (isZodV4Schema(schema) && zodToJsonSchemaV4) {
+        return normalizeJsonSchema(
+          zodToJsonSchemaV4(schema, { target: "draft-7" }),
+        );
+      }
+      if (zodToJsonSchemaV3) {
+        return normalizeJsonSchema(zodToJsonSchemaV3(schema));
+      }
+      if (zodToJsonSchemaV4) {
+        return normalizeJsonSchema(
+          zodToJsonSchemaV4(schema, { target: "draft-7" }),
+        );
+      }
+      return {};
+    } catch {
+      return {};
+    }
+  };
+}
+
+async function loadParameterSerializationHelpers(): Promise<ParameterSerializationHelpers> {
+  const braintrustPath = resolveBraintrustPath();
+  const zodToJsonSchema = loadZodSchemaSerializer(braintrustPath);
+  try {
+    const mod: unknown = await import(pathToFileURL(braintrustPath).href);
+    return {
+      sdkSerializeParameters: extractParameterSerializer(mod),
+      promptDefinitionToPromptData: extractPromptDefinitionSerializer(mod),
+      zodToJsonSchema,
+    };
+  } catch {
+    return {
+      sdkSerializeParameters: null,
+      promptDefinitionToPromptData: null,
+      zodToJsonSchema,
+    };
+  }
+}
+
+function extractParseParent(mod: unknown): ParseParentFunction | null {
+  if (!isObject(mod)) {
+    return null;
+  }
+  const candidate = Reflect.get(mod, "parseParent");
+  if (typeof candidate === "function") {
+    return candidate as ParseParentFunction;
+  }
+  const defaultExport = Reflect.get(mod, "default");
+  if (isObject(defaultExport)) {
+    const fromDefault = Reflect.get(defaultExport, "parseParent");
+    if (typeof fromDefault === "function") {
+      return fromDefault as ParseParentFunction;
+    }
+  }
+  return null;
+}
+
+function extractGlobalStateGetter(mod: unknown): (() => unknown) | null {
+  if (!isObject(mod)) {
+    return null;
+  }
+  const candidate = Reflect.get(mod, "_internalGetGlobalState");
+  if (typeof candidate === "function") {
+    return candidate as () => unknown;
+  }
+  const defaultExport = Reflect.get(mod, "default");
+  if (isObject(defaultExport)) {
+    const fromDefault = Reflect.get(defaultExport, "_internalGetGlobalState");
+    if (typeof fromDefault === "function") {
+      return fromDefault as () => unknown;
+    }
+  }
+  return null;
+}
+
+function loadBraintrustUtilParseParent(): ParseParentFunction | null {
+  const braintrustPath = resolveBraintrustPath();
+  const requireFromBraintrust = createRequire(
+    pathToFileURL(braintrustPath).href,
+  );
+  try {
+    const utilMod: unknown = requireFromBraintrust("braintrust/util");
+    return extractParseParent(utilMod);
+  } catch {
+    return null;
+  }
+}
+
+function propagateInheritedBraintrustState(braintrust: BraintrustModule) {
+  const getter = (braintrust as Record<string, unknown>)
+    ._internalGetGlobalState;
+  if (typeof getter !== "function") {
+    return;
+  }
+  const state = getter();
+  if (state !== undefined && state !== null) {
+    globalThis.__inherited_braintrust_state = state;
+  }
+}
+
+async function loadFiles(files: string[]): Promise<unknown[]> {
+  const modules: unknown[] = [];
+  // Internal CLI-controlled flag for ESM retry; not user-facing config.
+  const forceEsm = envFlag("BT_EVAL_FORCE_ESM");
+  // vite-node installs transform hooks that handle TypeScript (including
+  // extension-less imports) and CJS named-export interop natively. A failed
+  // require() corrupts Node's module cache and causes the subsequent import()
+  // to hit the "module imported again after being required" bug, so we skip
+  // require() for .ts/.tsx files entirely when running under vite-node.
+  const isViteNode = process.env.BT_EVAL_RUNNER_KIND === "vite-node";
+  for (const file of files) {
+    const fileUrl = pathToFileURL(file).href;
+    const isTypeScript = file.endsWith(".ts") || file.endsWith(".tsx");
+    const preferRequire =
+      !forceEsm &&
+      !(isViteNode && isTypeScript) &&
+      (isTypeScript || file.endsWith(".cjs"));
+
+    if (preferRequire) {
+      try {
+        const require = createRequire(fileUrl);
+        const mod = require(file);
+        modules.push(mod);
+        continue;
+      } catch (requireErr) {
+        try {
+          const mod = await import(fileUrl);
+          modules.push(mod);
+          continue;
+        } catch (esmErr) {
+          throw new Error(
+            `Failed to load ${file} as CJS (${formatError(requireErr)}) or ESM (${formatError(esmErr)}).`,
+          );
+        }
+      }
+    }
+
+    try {
+      const mod = await import(fileUrl);
+      modules.push(mod);
+      continue;
+    } catch (err) {
+      if (!shouldTryRequire(file, err)) {
+        throw err;
+      }
+      try {
+        const require = createRequire(fileUrl);
+        const mod = require(file);
+        modules.push(mod);
+        continue;
+      } catch (requireErr) {
+        throw new Error(
+          `Failed to load ${file} as ESM (${formatError(err)}) or CJS (${formatError(requireErr)}).`,
+        );
+      }
+    }
+  }
+  return modules;
+}
+
+function shouldTryRequire(file: string, err: unknown): boolean {
+  if (envFlag("BT_EVAL_FORCE_ESM")) {
+    return false;
+  }
+  if (process.env.BT_EVAL_RUNNER_KIND === "vite-node") {
+    return false;
+  }
+  if (process.env.BT_EVAL_CJS === "1" || file.endsWith(".cjs")) {
+    return true;
+  }
+  if (
+    (file.endsWith(".ts") || file.endsWith(".tsx")) &&
+    isNodeErrorCode(err, "ERR_UNKNOWN_FILE_EXTENSION")
+  ) {
+    return true;
+  }
+  if (!(err instanceof Error)) {
+    return false;
+  }
+  const message = err.message || "";
+  return (
+    message.includes("require is not defined") ||
+    message.includes("exports is not defined") ||
+    message.includes("module is not defined") ||
+    message.includes("Cannot use import statement outside a module")
+  );
+}
+
+function isNodeErrorCode(err: unknown, code: string): boolean {
+  if (!isObject(err) || !("code" in err)) {
+    return false;
+  }
+  return typeof err.code === "string" && err.code === code;
+}
+
+function formatError(err: unknown): string {
+  if (err instanceof Error) {
+    return err.message;
+  }
+  return String(err);
+}
+
+function createEvalProgressReporter(
+  sse: SseWriter | null,
+  evaluatorName: string,
+) {
+  let activeName = evaluatorName;
+  return {
+    start: (name: string, total: number) => {
+      activeName = name;
+      sendEvalProgress(sse, name, "start", total);
+    },
+    stop: () => {
+      if (activeName) {
+        sendEvalProgress(sse, activeName, "stop");
+      }
+    },
+    increment: (name: string) => {
+      sendEvalProgress(sse, name, "increment");
+    },
+    setTotal: (name: string, total: number) => {
+      sendEvalProgress(sse, name, "set_total", total);
+    },
+  };
+}
+
+function sendEvalProgress(
+  sse: SseWriter | null,
+  evaluatorName: string,
+  kind: "start" | "increment" | "set_total" | "stop",
+  total?: number,
+) {
+  if (!sse) {
+    return;
+  }
+  sse.send("progress", {
+    id: `eval-progress:${evaluatorName}`,
+    object_type: "task",
+    format: "global",
+    output_type: "any",
+    name: evaluatorName,
+    event: "progress",
+    data: JSON.stringify({
+      type: "eval_progress",
+      kind,
+      ...(total !== undefined ? { total } : {}),
+    }),
+  });
+}
+
+function serializeError(err: unknown) {
+  if (err instanceof Error) {
+    return { message: err.message, stack: err.stack };
+  }
+  return { message: String(err) };
+}
+
+function sendConsole(
+  sse: SseWriter | null,
+  message: string,
+  stream: "stdout" | "stderr" = "stderr",
+) {
+  if (!sse) {
+    if (stream === "stderr") {
+      console.error(message);
+    } else {
+      console.log(message);
+    }
+    return;
+  }
+  sse.send("console", { stream, message });
+}
+
+function getEvaluators(): EvaluatorEntry[] {
+  const evals = globalThis._evals;
+  if (!evals || !evals.evaluators) {
+    return [];
+  }
+  return Object.values(evals.evaluators) as EvaluatorEntry[];
+}
+
+function getReporters(): Record<string, unknown> {
+  const evals = globalThis._evals;
+  if (!evals || !evals.reporters) {
+    return {};
+  }
+  return evals.reporters as Record<string, unknown>;
+}
+
+function resolveReporter(
+  reporter: unknown,
+  reporters: Record<string, unknown>,
+): unknown | undefined {
+  if (typeof reporter === "string") {
+    if (!(reporter in reporters)) {
+      throw new Error(`Reporter ${reporter} not found`);
+    }
+    return reporters[reporter];
+  }
+  if (reporter !== undefined && reporter !== null) {
+    return reporter;
+  }
+
+  const values = Object.values(reporters);
+  if (values.length === 0) {
+    return undefined;
+  }
+  if (values.length === 1) {
+    return values[0];
+  }
+  const names = Object.keys(reporters).join(", ");
+  throw new Error(
+    `Multiple reporters found (${names}). Please specify a reporter explicitly.`,
+  );
+}
+
+function evaluateFilter(
+  object: Record<string, unknown>,
+  filter: EvalFilter,
+): boolean {
+  const key = filter.path.reduce<unknown>((acc, part) => {
+    if (!isObject(acc)) {
+      return undefined;
+    }
+    return acc[part];
+  }, object);
+  if (key === undefined) {
+    return false;
+  }
+  return filter.pattern.test(serializeJSONWithPlainString(key));
+}
+
+function filterEvaluators(
+  evaluators: EvaluatorEntry[],
+  filters: EvalFilter[],
+): EvaluatorEntry[] {
+  if (filters.length === 0) {
+    return evaluators;
+  }
+  return evaluators.filter((entry) =>
+    filters.every((filter) => evaluateFilter(entry.evaluator, filter)),
+  );
+}
+
+function extractScoreName(score: unknown, idx: number): string {
+  if (typeof score === "function" && typeof score.name === "string") {
+    return score.name || `scorer_${idx}`;
+  }
+  return `scorer_${idx}`;
+}
+
+async function serializeEvaluatorParameters(
+  raw: unknown,
+  helpers?: ParameterSerializationHelpers,
+): Promise<unknown | undefined> {
+  if (raw === undefined || raw === null) {
+    return undefined;
+  }
+
+  const resolved = raw instanceof Promise ? await raw : raw;
+  if (!isObject(resolved)) {
+    return undefined;
+  }
+
+  if (helpers?.sdkSerializeParameters) {
+    try {
+      return helpers.sdkSerializeParameters(resolved);
+    } catch {
+      // Fallback to legacy serialization below when SDK internals are unavailable.
+    }
+  }
+
+  const marker = Reflect.get(resolved, "__braintrust_parameters_marker");
+  if (marker === true) {
+    const schema = Reflect.get(resolved, "schema");
+    const source = {
+      parametersId: Reflect.get(resolved, "id"),
+      slug: Reflect.get(resolved, "slug"),
+      name: Reflect.get(resolved, "name"),
+      projectId: Reflect.get(resolved, "projectId"),
+      version: Reflect.get(resolved, "version"),
+    };
+    return {
+      type: "braintrust.parameters",
+      schema,
+      source,
+    };
+  }
+
+  const schema: Record<string, unknown> = {};
+  for (const [name, value] of Object.entries(resolved)) {
+    if (isObject(value) && value.type === "prompt") {
+      let promptDefault = value.default;
+      if (
+        promptDefault !== undefined &&
+        helpers?.promptDefinitionToPromptData
+      ) {
+        try {
+          promptDefault = helpers.promptDefinitionToPromptData(promptDefault);
+        } catch {
+          // Keep raw prompt default when conversion utility is unavailable.
+        }
+      }
+      schema[name] = {
+        type: "prompt",
+        ...(promptDefault !== undefined ? { default: promptDefault } : {}),
+        ...(typeof value.description === "string"
+          ? { description: value.description }
+          : {}),
+      };
+    } else {
+      const jsonSchema = helpers?.zodToJsonSchema
+        ? helpers.zodToJsonSchema(value)
+        : {};
+      schema[name] = {
+        type: "data",
+        schema: jsonSchema,
+        ...(Object.prototype.hasOwnProperty.call(jsonSchema, "default")
+          ? { default: Reflect.get(jsonSchema, "default") }
+          : {}),
+        ...(typeof Reflect.get(jsonSchema, "description") === "string"
+          ? { description: Reflect.get(jsonSchema, "description") as string }
+          : {}),
+      };
+    }
+  }
+
+  return {
+    type: "braintrust.staticParameters",
+    schema,
+    source: null,
+  };
+}
+
+async function buildEvaluatorDefinitions(evaluators: EvaluatorEntry[]) {
+  const serializationHelpers = await loadParameterSerializationHelpers();
+  const result: Record<
+    string,
+    { parameters?: unknown; scores: Array<{ name: string }> }
+  > = {};
+
+  for (const entry of evaluators) {
+    const scores = Array.isArray(entry.evaluator.scores)
+      ? entry.evaluator.scores.map((score, idx) => ({
+          name: extractScoreName(score, idx),
+        }))
+      : [];
+
+    const parameters = await serializeEvaluatorParameters(
+      entry.evaluator.parameters,
+      serializationHelpers,
+    );
+
+    result[entry.evaluator.evalName] = {
+      ...(parameters !== undefined ? { parameters } : {}),
+      scores,
+    };
+  }
+
+  return result;
+}
+
+function parseEvalRequest(raw: string | null): EvalRequest {
+  if (!raw) {
+    throw new Error("Missing BT_EVAL_DEV_REQUEST_JSON");
+  }
+  try {
+    const parsed = JSON.parse(raw);
+    if (!isObject(parsed)) {
+      throw new Error("BT_EVAL_DEV_REQUEST_JSON must be a JSON object.");
+    }
+    if (typeof parsed.name !== "string" || parsed.name.length === 0) {
+      throw new Error("Eval request must include a non-empty name.");
+    }
+    if (!isObject(parsed.data)) {
+      throw new Error("Eval request must include a data object.");
+    }
+    return parsed as EvalRequest;
+  } catch (err) {
+    if (err instanceof Error) {
+      throw err;
+    }
+    throw new Error(String(err));
+  }
+}
+
+function normalizeParent(
+  parent: unknown,
+  parseParent: ParseParentFunction | null,
+): string | undefined {
+  if (parent === undefined || parent === null) {
+    return undefined;
+  }
+  if (typeof parent === "string") {
+    return parent;
+  }
+  if (parseParent) {
+    try {
+      const parsed = parseParent(parent);
+      if (typeof parsed === "string" || parsed === undefined) {
+        return parsed;
+      }
+    } catch {
+      // Fall through and return a validation error below.
+    }
+  }
+  throw new Error(
+    "Invalid parent payload: expected a serialized parent string or a valid parent object.",
+  );
+}
+
+function resolveEvalData(
+  data: EvalRequest["data"],
+  initDataset: InitDatasetFunction | undefined,
+): unknown {
+  if ("data" in data) {
+    return data.data;
+  }
+  if ("project_name" in data && "dataset_name" in data) {
+    if (!initDataset) {
+      throw new Error(
+        "Unable to resolve dataset references: initDataset() is unavailable.",
+      );
+    }
+    return initDataset(data.project_name, {
+      dataset: data.dataset_name,
+      _internal_btql: data._internal_btql,
+    });
+  }
+  if ("project_id" in data && "dataset_name" in data) {
+    if (!initDataset) {
+      throw new Error(
+        "Unable to resolve dataset references: initDataset() is unavailable.",
+      );
+    }
+    return initDataset({
+      projectId: data.project_id,
+      dataset: data.dataset_name,
+      _internal_btql: data._internal_btql,
+    });
+  }
+  throw new Error("Invalid eval data payload.");
+}
+
+type SamplingMetadata = {
+  runMode: "full" | "first" | "sample";
+  isFinal: boolean;
+  runLabel: string;
+  sampleCount?: number;
+  sampleSeed?: number;
+};
+
+function samplingMetadata(config: RunnerConfig): SamplingMetadata {
+  if (config.first !== null) {
+    return {
+      runMode: "first",
+      isFinal: false,
+      runLabel: `Run mode: first ${config.first} examples (non-final smoke run)`,
+      sampleCount: config.first,
+    };
+  }
+  if (config.sample !== null) {
+    const seed = config.sampleSeed ?? 0;
+    return {
+      runMode: "sample",
+      isFinal: false,
+      runLabel: `Run mode: random sample of ${config.sample} examples (seed ${seed}, non-final smoke run)`,
+      sampleCount: config.sample,
+      sampleSeed: seed,
+    };
+  }
+  return {
+    runMode: "full",
+    isFinal: true,
+    runLabel: "Run mode: full dataset",
+  };
+}
+
+function attachSamplingSummary(
+  summary: unknown,
+  config: RunnerConfig,
+): unknown {
+  const metadata = samplingMetadata(config);
+  if (isObject(summary)) {
+    return {
+      ...summary,
+      runMode: metadata.runMode,
+      isFinal: metadata.isFinal,
+      runLabel: metadata.runLabel,
+      ...(metadata.sampleCount !== undefined
+        ? { sampleCount: metadata.sampleCount }
+        : {}),
+      ...(metadata.sampleSeed !== undefined
+        ? { sampleSeed: metadata.sampleSeed }
+        : {}),
+    };
+  }
+  return {
+    summary,
+    runMode: metadata.runMode,
+    isFinal: metadata.isFinal,
+    runLabel: metadata.runLabel,
+    ...(metadata.sampleCount !== undefined
+      ? { sampleCount: metadata.sampleCount }
+      : {}),
+    ...(metadata.sampleSeed !== undefined
+      ? { sampleSeed: metadata.sampleSeed }
+      : {}),
+  };
+}
+
+function isPromiseLike(value: unknown): value is PromiseLike<unknown> {
+  return (
+    typeof value === "object" &&
+    value !== null &&
+    typeof Reflect.get(value, "then") === "function"
+  );
+}
+
+function isAsyncIterable(value: unknown): value is AsyncIterable<unknown> {
+  return (
+    typeof value === "object" &&
+    value !== null &&
+    typeof Reflect.get(value, Symbol.asyncIterator) === "function"
+  );
+}
+
+function isIterable(value: unknown): value is Iterable<unknown> {
+  return (
+    typeof value === "object" &&
+    value !== null &&
+    typeof Reflect.get(value, Symbol.iterator) === "function"
+  );
+}
+
+function isDatasetLike(value: unknown): value is {
+  fetch: (options?: { batchSize?: number }) => AsyncIterable<unknown>;
+} {
+  return (
+    typeof value === "object" &&
+    value !== null &&
+    typeof Reflect.get(value, "fetch") === "function" &&
+    typeof Reflect.get(value, "summarize") === "function"
+  );
+}
+
+function createSeededRandom(seed: number): () => number {
+  // SplitMix64 keeps seed entropy beyond 32 bits so distinct seeds stay distinct.
+  let state = BigInt.asUintN(64, BigInt(seed));
+  return () => {
+    state = BigInt.asUintN(64, state + 0x9e3779b97f4a7c15n);
+    let z = state;
+    z = BigInt.asUintN(64, (z ^ (z >> 30n)) * 0xbf58476d1ce4e5b9n);
+    z = BigInt.asUintN(64, (z ^ (z >> 27n)) * 0x94d049bb133111ebn);
+    z = BigInt.asUintN(64, z ^ (z >> 31n));
+    return Number(z & 0x1fffffffffffffn) / 9007199254740992;
+  };
+}
+
+type SamplingSourceOptions = {
+  initialCallReceiver?: unknown;
+};
+
+async function resolveSamplingSource(
+  source: unknown,
+  options?: SamplingSourceOptions,
+): Promise<unknown> {
+  let current: unknown = source;
+  let firstCall = true;
+  while (true) {
+    current = isPromiseLike(current) ? await current : current;
+    if (typeof current !== "function") {
+      return current;
+    }
+    if (firstCall && options?.initialCallReceiver !== undefined) {
+      current = Reflect.apply(
+        current as (...args: unknown[]) => unknown,
+        options.initialCallReceiver,
+        [],
+      );
+      firstCall = false;
+      continue;
+    }
+    current = (current as () => unknown)();
+    firstCall = false;
+  }
+}
+
+async function* iterateDataSource(
+  source: unknown,
+  batchSizeHint?: number,
+  options?: SamplingSourceOptions,
+): AsyncGenerator<unknown> {
+  const resolved = await resolveSamplingSource(source, options);
+  if (Array.isArray(resolved)) {
+    for (const item of resolved) {
+      yield item;
+    }
+    return;
+  }
+  if (isDatasetLike(resolved)) {
+    const options =
+      batchSizeHint !== undefined ? { batchSize: batchSizeHint } : undefined;
+    for await (const item of resolved.fetch(options)) {
+      yield item;
+    }
+    return;
+  }
+  if (isAsyncIterable(resolved)) {
+    for await (const item of resolved) {
+      yield item;
+    }
+    return;
+  }
+  if (typeof resolved !== "string" && isIterable(resolved)) {
+    for (const item of resolved) {
+      yield item;
+    }
+    return;
+  }
+  throw new Error(
+    "Sampling is only supported for arrays, iterables, async iterables, and Braintrust datasets.",
+  );
+}
+
+async function collectFirstRecords(
+  source: unknown,
+  count: number,
+  options?: SamplingSourceOptions,
+): Promise<unknown[]> {
+  const items: unknown[] = [];
+  for await (const item of iterateDataSource(source, count, options)) {
+    items.push(item);
+    if (items.length >= count) {
+      break;
+    }
+  }
+  return items;
+}
+
+async function reservoirSampleRecords(
+  source: unknown,
+  count: number,
+  seed: number,
+  options?: SamplingSourceOptions,
+): Promise<unknown[]> {
+  const random = createSeededRandom(seed);
+  const sample: unknown[] = [];
+  let seen = 0;
+  for await (const item of iterateDataSource(source, undefined, options)) {
+    seen += 1;
+    if (sample.length < count) {
+      sample.push(item);
+      continue;
+    }
+    const index = Math.floor(random() * seen);
+    if (index < count) {
+      sample[index] = item;
+    }
+  }
+  return sample;
+}
+
+async function applySamplingToData(
+  data: unknown,
+  config: RunnerConfig,
+  options?: SamplingSourceOptions,
+): Promise<unknown> {
+  if (config.first !== null) {
+    return await collectFirstRecords(data, config.first, options);
+  }
+  if (config.sample !== null) {
+    return await reservoirSampleRecords(
+      data,
+      config.sample,
+      config.sampleSeed ?? 0,
+      options,
+    );
+  }
+  return data;
+}
+
+function convertFunctionId(
+  functionId: Record<string, unknown>,
+): Record<string, unknown> {
+  const converted: Record<string, unknown> = {};
+  if (functionId.function_id !== undefined) {
+    converted.function_id = functionId.function_id;
+  }
+  if (typeof functionId.project_name === "string") {
+    converted.projectName = functionId.project_name;
+  }
+  if (typeof functionId.slug === "string") {
+    converted.slug = functionId.slug;
+  }
+  if (typeof functionId.global_function === "string") {
+    converted.globalFunction = functionId.global_function;
+  }
+  if (typeof functionId.function_type === "string") {
+    converted.functionType = functionId.function_type;
+  }
+  if (typeof functionId.prompt_session_id === "string") {
+    converted.promptSessionId = functionId.prompt_session_id;
+  }
+  if (typeof functionId.prompt_session_function_id === "string") {
+    converted.promptSessionFunctionId = functionId.prompt_session_function_id;
+  }
+  if (typeof functionId.version === "string") {
+    converted.version = functionId.version;
+  }
+  return converted;
+}
+
+function makeEvalScorer(
+  invoke: InvokeFunction,
+  score: EvalScoreSpec,
+  projectId: string | undefined,
+) {
+  const scorer = async (args: Record<string, unknown>) => {
+    return await invoke({
+      ...convertFunctionId(score.function_id),
+      input: {
+        input: args.input,
+        output: args.output,
+        expected: args.expected,
+        metadata: args.metadata,
+      },
+      stream: false,
+      mode: "auto",
+      strict: true,
+      ...(projectId ? { projectId } : {}),
+    });
+  };
+
+  Object.defineProperty(scorer, "name", {
+    value: score.name,
+    writable: false,
+  });
+
+  return scorer;
+}
+
+function inferEvalErrorStatus(err: unknown): number {
+  if (!(err instanceof Error)) {
+    return 500;
+  }
+  const message = err.message.toLowerCase();
+  if (
+    message.includes("invalid parameter") ||
+    message.includes("invalid parameters") ||
+    message.includes("must include") ||
+    message.includes("invalid eval")
+  ) {
+    return 400;
+  }
+  if (message.includes("not found")) {
+    return 404;
+  }
+  return 500;
+}
+
+function sendEvalError(sse: SseWriter | null, err: unknown, status?: number) {
+  const payload =
+    err instanceof Error
+      ? {
+          message: err.message,
+          stack: err.stack,
+          ...(status !== undefined ? { status } : {}),
+        }
+      : {
+          message: String(err),
+          ...(status !== undefined ? { status } : {}),
+        };
+  if (sse) {
+    sse.send("error", payload);
+  } else {
+    console.error(payload.message);
+  }
+}
+
+async function runRequestedEval(config: RunnerConfig, runner: EvalRunner) {
+  let request: EvalRequest;
+  try {
+    request = parseEvalRequest(config.devRequestJson);
+  } catch (err) {
+    sendEvalError(runner.sse, err, 400);
+    return;
+  }
+
+  const entry = getEvaluators().find(
+    (candidate) => candidate.evaluator.evalName === request.name,
+  );
+  if (!entry) {
+    sendEvalError(
+      runner.sse,
+      new Error(`Evaluator '${request.name}' not found`),
+      404,
+    );
+    return;
+  }
+
+  try {
+    const baseScores = Array.isArray(entry.evaluator.scores)
+      ? entry.evaluator.scores
+      : [];
+
+    if ((request.scores?.length ?? 0) > 0 && !runner.invoke) {
+      throw new Error(
+        "Unable to run scorer functions: invoke() is unavailable.",
+      );
+    }
+
+    const extraScores = (request.scores ?? []).map((score) =>
+      makeEvalScorer(
+        runner.invoke as InvokeFunction,
+        score,
+        request.project_id,
+      ),
+    );
+
+    const state = runner.getState ? runner.getState() : undefined;
+    const evaluator = {
+      ...entry.evaluator,
+      data: resolveEvalData(request.data, runner.initDataset),
+      scores: baseScores.concat(extraScores),
+      ...(request.experiment_name
+        ? { experimentName: request.experiment_name }
+        : {}),
+      ...(request.project_id ? { projectId: request.project_id } : {}),
+      ...(state !== undefined && state !== null ? { state } : {}),
+    };
+
+    const options: EvalOptions = {};
+    if (request.parameters) {
+      options.parameters = request.parameters;
+    }
+    if (request.parent !== undefined) {
+      options.parent = normalizeParent(request.parent, runner.parseParent);
+    }
+
+    const reporters = getReporters();
+    const resolvedReporter = resolveReporter(entry.reporter, reporters);
+    if (resolvedReporter !== undefined) {
+      options.reporter = resolvedReporter;
+    }
+
+    await runner.runEval(
+      entry.evaluator.projectName,
+      evaluator,
+      Object.keys(options).length > 0 ? options : undefined,
+    );
+  } catch (err) {
+    sendEvalError(runner.sse, err, inferEvalErrorStatus(err));
+  }
+}
+
+function extractBtEvalMain(mod: unknown): BtEvalMain | null {
+  if (!mod || typeof mod !== "object") {
+    return null;
+  }
+  const candidate = mod as Record<string, unknown>;
+  if (typeof candidate.btEvalMain === "function") {
+    return candidate.btEvalMain as BtEvalMain;
+  }
+  const defaultExport = candidate.default as
+    | Record<string, unknown>
+    | undefined;
+  if (defaultExport && typeof defaultExport.btEvalMain === "function") {
+    return defaultExport.btEvalMain as BtEvalMain;
+  }
+  return null;
+}
+
+function collectBtEvalMains(mods: unknown[]): BtEvalMain[] {
+  const mains: BtEvalMain[] = [];
+  for (const mod of mods) {
+    const main = extractBtEvalMain(mod);
+    if (main) {
+      mains.push(main);
+    }
+  }
+  return mains;
+}
+
+function shouldDisableSendLogs(): boolean {
+  return (
+    process.env.BT_EVAL_NO_SEND_LOGS === "1" ||
+    process.env.BT_EVAL_LOCAL === "1"
+  );
+}
+
+function getEvaluatorName(
+  evaluator: Record<string, unknown>,
+  fallback: string,
+): string {
+  const candidate = evaluator.evalName ?? evaluator.name ?? evaluator.task;
+  if (typeof candidate === "string" && candidate.length > 0) {
+    return candidate;
+  }
+  return fallback;
+}
+
+function wrapTaskForStreamingProgress(
+  evaluator: Record<string, unknown>,
+): Record<string, unknown> {
+  const task = evaluator.task;
+  if (typeof task !== "function") {
+    return evaluator;
+  }
+
+  const wrappedTask = async (input: unknown, hooks: unknown) => {
+    const result = await task(input, hooks);
+    if (isObject(hooks)) {
+      const reportProgress = Reflect.get(hooks, "reportProgress");
+      if (typeof reportProgress === "function") {
+        try {
+          reportProgress({
+            format: "code",
+            output_type: "completion",
+            event: "json_delta",
+            data: JSON.stringify(result),
+          });
+        } catch {
+          // Progress updates should not fail the evaluator run.
+        }
+      }
+    }
+    return result;
+  };
+
+  return {
+    ...evaluator,
+    task: wrappedTask,
+  };
+}
+
+function mergeEvalOptions(
+  base: EvalOptions,
+  overrides?: EvalOptions,
+): EvalOptions {
+  if (!overrides) {
+    return base;
+  }
+
+  const merged: EvalOptions = { ...base, ...overrides };
+
+  const baseProgress = base.progress as Record<string, unknown> | undefined;
+  const overrideProgress = overrides.progress as
+    | Record<string, unknown>
+    | undefined;
+  if (baseProgress || overrideProgress) {
+    merged.progress = mergeProgress(baseProgress, overrideProgress);
+  }
+
+  const baseStream = base.stream as ((data: unknown) => void) | undefined;
+  const overrideStream = overrides.stream as
+    | ((data: unknown) => void)
+    | undefined;
+  if (baseStream || overrideStream) {
+    merged.stream = mergeHandlers(baseStream, overrideStream);
+  }
+
+  const baseOnStart = base.onStart as ((data: unknown) => void) | undefined;
+  const overrideOnStart = overrides.onStart as
+    | ((data: unknown) => void)
+    | undefined;
+  if (baseOnStart || overrideOnStart) {
+    merged.onStart = mergeHandlers(baseOnStart, overrideOnStart);
+  }
+
+  if (base.reporter && overrides.reporter === undefined) {
+    merged.reporter = base.reporter;
+  }
+
+  return merged;
+}
+
+function mergeHandlers<Args extends unknown[]>(
+  base?: (...args: Args) => void,
+  override?: (...args: Args) => void,
+): ((...args: Args) => void) | undefined {
+  if (base && override) {
+    return (...args: Args) => {
+      base(...args);
+      override(...args);
+    };
+  }
+  return base ?? override;
+}
+
+function mergeProgress(
+  base?: Partial<ProgressReporter>,
+  override?: Partial<ProgressReporter>,
+): ProgressReporter | undefined {
+  if (!base) {
+    return override as ProgressReporter | undefined;
+  }
+  if (!override) {
+    return base as ProgressReporter;
+  }
+  const noopName = (_name: string) => {};
+  const noopStart = (_name: string, _total: number) => {};
+  return {
+    start:
+      mergeHandlers(base.start, override.start) ??
+      base.start ??
+      override.start ??
+      noopStart,
+    stop:
+      mergeHandlers(base.stop, override.stop) ??
+      base.stop ??
+      override.stop ??
+      noopName,
+    increment:
+      mergeHandlers(base.increment, override.increment) ??
+      base.increment ??
+      override.increment ??
+      noopName,
+    setTotal:
+      mergeHandlers(base.setTotal, override.setTotal) ??
+      base.setTotal ??
+      override.setTotal ??
+      noopStart,
+  };
+}
+
+async function createEvalRunner(config: RunnerConfig): Promise<EvalRunner> {
+  const braintrust = await loadBraintrust();
+  const Eval = braintrust.Eval;
+  if (typeof Eval !== "function") {
+    throw new Error("Unable to load Eval() from braintrust package.");
+  }
+  const login = braintrust.login;
+  const initDataset = braintrust.initDataset;
+  const invoke = braintrust.invoke;
+
+  const sse = createSseWriter();
+  const noSendLogs = shouldDisableSendLogs();
+  const parseParent = loadBraintrustUtilParseParent();
+  const getState = extractGlobalStateGetter(braintrust);
+
+  const makeEvalOptions = (
+    evaluatorName: string,
+    overrides?: EvalOptions,
+  ): EvalOptions | undefined => {
+    let base: EvalOptions = {};
+    if (noSendLogs) {
+      base.noSendLogs = true;
+    }
+    if (sse) {
+      base = {
+        ...base,
+        reporter: {
+          name: "bt-silent-reporter",
+          reportEval: () => true,
+          reportRun: () => true,
+        },
+        progress: createEvalProgressReporter(sse, evaluatorName),
+        stream: (data: unknown) => {
+          sse.send("progress", data);
+        },
+        onStart: (metadata: unknown) => {
+          sse.send("start", metadata);
+        },
+      };
+    }
+
+    if (!overrides) {
+      return Object.keys(base).length === 0 ? undefined : base;
+    }
+    return mergeEvalOptions(base, overrides);
+  };
+
+  const runEval = async (
+    projectName: string,
+    evaluator: Record<string, unknown>,
+    options?: EvalOptions,
+    paramsOverride?: Record<string, unknown>,
+  ) => {
+    globalThis._lazy_load = false;
+    const evaluatorName = getEvaluatorName(evaluator, projectName);
+    // Only inject CLI params when the evaluator declares a parameters schema.
+    // Drop any --param keys the evaluator doesn't declare so a single command
+    // running multiple evaluators doesn't fail on unrelated params.
+    const effectiveParams =
+      paramsOverride != null
+        ? { ...(config.params ?? {}), ...paramsOverride }
+        : config.params;
+    const filteredParams =
+      effectiveParams != null && evaluator.parameters != null
+        ? filterParamsForEvaluator(effectiveParams, evaluator.parameters)
+        : null;
+    const effectiveOptions =
+      filteredParams != null
+        ? mergeEvalOptions({ parameters: filteredParams }, options)
+        : options;
+    const opts = makeEvalOptions(evaluatorName, effectiveOptions);
+    const sampledData = await applySamplingToData(evaluator.data, config, {
+      initialCallReceiver: evaluator,
+    });
+    const wrappedEvaluator = wrapTaskForStreamingProgress({
+      ...evaluator,
+      data: sampledData,
+    });
+    const result = await Eval(projectName, wrappedEvaluator, opts);
+    const summary = attachSamplingSummary(result.summary, config);
+    const failingResults = result.results.filter(
+      (r: { error?: unknown }) => r.error !== undefined,
+    );
+    if (failingResults.length > 0 && sse) {
+      sendConsole(
+        sse,
+        `Evaluator ${evaluatorName} failed with ${failingResults.length} error${failingResults.length === 1 ? "" : "s"}.`,
+      );
+    }
+    if (sse) {
+      sse.send("summary", summary);
+    } else if (config.jsonl) {
+      console.log(JSON.stringify(summary));
+    }
+    return result;
+  };
+
+  const runRegisteredEvals = async (evaluators: EvaluatorEntry[]) => {
+    if (sse) {
+      sse.send("processing", { evaluators: evaluators.length });
+    }
+    const reporters = getReporters();
+    const runEntry = async (entry: EvaluatorEntry): Promise<boolean> => {
+      try {
+        const resolvedReporter = resolveReporter(entry.reporter, reporters);
+        const options =
+          resolvedReporter !== undefined
+            ? { reporter: resolvedReporter }
+            : undefined;
+        const result = await runEval(
+          entry.evaluator.projectName,
+          entry.evaluator,
+          options,
+          entry.paramsOverride,
+        );
+        const failingResults = result.results.filter(
+          (r: { error?: unknown }) => r.error !== undefined,
+        );
+        if (failingResults.length > 0 && resolvedReporter === undefined) {
+          return false;
+        }
+        return true;
+      } catch (err) {
+        if (sse) {
+          sse.send("error", serializeError(err));
+        } else {
+          console.error(err);
+        }
+        return false;
+      }
+    };
+
+    if (config.terminateOnFailure) {
+      for (const entry of evaluators) {
+        const ok = await runEntry(entry);
+        if (!ok) {
+          return false;
+        }
+      }
+      return true;
+    }
+
+    const results = await Promise.all(
+      evaluators.map((entry) => runEntry(entry)),
+    );
+    return results.every(Boolean);
+  };
+
+  const finish = (ok: boolean) => {
+    if (sse) {
+      sse.send("dependencies", { files: collectDependencyFiles() });
+      sse.send("done", "");
+      sse.close();
+    }
+    if (!ok) {
+      process.exitCode = 1;
+    }
+  };
+
+  return {
+    Eval,
+    sse,
+    login,
+    initDataset,
+    invoke,
+    runEval,
+    runRegisteredEvals,
+    makeEvalOptions,
+    finish,
+    noSendLogs,
+    parseParent,
+    getState,
+  };
+}
+
+export async function main() {
+  const config = readRunnerConfig();
+  const files = process.argv.slice(2);
+  if (files.length === 0) {
+    console.error("No eval files provided.");
+    process.exit(1);
+  }
+
+  installDependencyTracking();
+  const normalized = normalizeFiles(files);
+  for (const file of normalized) {
+    maybeRecordDependency(file);
+  }
+  collectStaticLocalDependencies(normalized);
+  ensureBraintrustAvailable();
+  const braintrust = await loadBraintrust();
+  propagateInheritedBraintrustState(braintrust);
+  initRegistry();
+  // Replace process.argv with [runtime, script, ...extraArgs] so that user
+  // code sees only the args they explicitly passed via `bt eval file -- ...`
+  // and not the eval file paths the runner uses internally.
+  const extraArgs: string[] = process.env.BT_EVAL_EXTRA_ARGS_JSON
+    ? (JSON.parse(process.env.BT_EVAL_EXTRA_ARGS_JSON) as string[])
+    : [];
+  process.argv = [...process.argv.slice(0, 2), ...extraArgs];
+  const modules = await loadFiles(normalized);
+  const btEvalMains = collectBtEvalMains(modules);
+
+  const runner = await createEvalRunner(config);
+  if (!runner.noSendLogs && typeof runner.login === "function") {
+    try {
+      await runner.login({});
+    } catch (err) {
+      if (runner.sse) {
+        runner.sse.send("error", serializeError(err));
+      } else {
+        console.error(err);
+      }
+      runner.finish(false);
+      return;
+    }
+  }
+  const context: BtEvalContext = {
+    Eval: runner.Eval,
+    runEval: runner.runEval,
+    runRegisteredEvals: () =>
+      runner.runRegisteredEvals(
+        filterEvaluators(getEvaluators(), config.filters),
+      ),
+    makeEvalOptions: runner.makeEvalOptions,
+    sendConsole: (message: string, stream?: "stdout" | "stderr") => {
+      sendConsole(runner.sse, message, stream);
+    },
+    sendEvent: (event: string, data: unknown) => {
+      if (runner.sse) {
+        runner.sse.send(event, data);
+      }
+    },
+  };
+
+  let ok = true;
+  try {
+    const discoveredEvaluators = getEvaluators();
+    let filteredEvaluators = filterEvaluators(
+      discoveredEvaluators,
+      config.filters,
+    );
+
+    if (config.devMode === "list") {
+      const definitions = await buildEvaluatorDefinitions(filteredEvaluators);
+      console.log(JSON.stringify(definitions));
+      return;
+    }
+
+    if (config.devMode === "eval") {
+      await runRequestedEval(config, runner);
+      return;
+    }
+
+    if (config.list) {
+      for (const entry of filteredEvaluators) {
+        console.log(entry.evaluator.evalName);
+      }
+      return;
+    }
+
+    if (btEvalMains.length > 0) {
+      if (config.matrix && config.matrix.length > 0) {
+        const message =
+          "--matrix-param is not supported for eval files that export btEvalMain. Remove the btEvalMain export or drop --matrix-param.";
+        if (runner.sse) {
+          runner.sse.send("error", serializeError(new Error(message)));
+        } else {
+          console.error(message);
+        }
+        ok = false;
+        return;
+      }
+      globalThis._lazy_load = false;
+      for (const main of btEvalMains) {
+        try {
+          await main(context);
+        } catch (err) {
+          ok = false;
+          if (runner.sse) {
+            runner.sse.send("error", serializeError(err));
+          } else {
+            console.error(err);
+          }
+        }
+      }
+    } else {
+      if (config.matrix && config.matrix.length > 0) {
+        if (filteredEvaluators.length !== 1) {
+          const names = filteredEvaluators.map((e) => e.evaluator.evalName);
+          const listed =
+            names.length === 0
+              ? "  (none matched)"
+              : names.map((n) => `  - ${n}`).join("\n");
+          const message = `--matrix-param is not supported when running multiple evals.\nMatched evals:\n${listed}\nUse --filter to select exactly one eval.`;
+          if (runner.sse) {
+            runner.sse.send("error", serializeError(new Error(message)));
+          } else {
+            console.error(message);
+          }
+          ok = false;
+          return;
+        }
+        const [sourceEntry] = filteredEvaluators;
+        const combos = matrixCombinations(config.matrix);
+        filteredEvaluators = combos.map((combo) => {
+          const label = formatMatrixLabel(combo);
+          const paramsOverride: Record<string, unknown> = {};
+          for (const [key, value] of combo) {
+            paramsOverride[key] = value;
+          }
+          const origExperimentName = sourceEntry.evaluator.experimentName;
+          const newExperimentName =
+            typeof origExperimentName === "string" &&
+            origExperimentName.length > 0
+              ? `${origExperimentName} [${label}]`
+              : `[${label}]`;
+          // Disambiguate the progress-bar / tracking key too — the runner emits progress
+          // events keyed by evalName, so combos must have distinct evalNames or bars collapse.
+          const newEvalName = `${sourceEntry.evaluator.evalName} [${label}]`;
+          return {
+            evaluator: {
+              ...sourceEntry.evaluator,
+              evalName: newEvalName,
+              experimentName: newExperimentName,
+            },
+            reporter: sourceEntry.reporter,
+            paramsOverride,
+          } satisfies EvaluatorEntry;
+        });
+      }
+      if (discoveredEvaluators.length === 0) {
+        console.error("No evaluators found. Did you call Eval() in the file?");
+        process.exit(1);
+      }
+      ok = await runner.runRegisteredEvals(filteredEvaluators);
+    }
+  } finally {
+    collectRequireCacheDependencies();
+    await collectDenoInfoDependencies(normalized);
+    runner.finish(ok);
+  }
+}
