+++ tests/eval_fixtures.rs
@@ -358,7 +358,19 @@ fn eval_runner_list_mode_serializes_parameter_defaults() {
         );
     }
 
-    let runner_script = root.join("scripts").join("eval-runner.mts");
+    // The default `.ts` runner is generated by bt at materialization time
+    // (impl body + appended fire-and-forget). Reproduce that here so we can
+    // hand a standalone script to ts-node without depending on the in-cwd
+    // cache from a prior bt run.
+    let impl_source = fs::read_to_string(root.join("scripts").join("eval-runner-impl.ts"))
+        .expect("read eval-runner-impl.ts");
+    let standalone = format!(
+        "{impl_source}\nmain().catch((err) => {{\n  console.error(err);\n  process.exit(1);\n}});\n"
+    );
+    let runner_script = fixture_dir.join(".bt").join("test-eval-runner.ts");
+    fs::create_dir_all(runner_script.parent().expect("parent")).expect("create .bt dir");
+    fs::write(&runner_script, &standalone).expect("write standalone runner");
+
     let output = Command::new(&runner)
         .arg(&runner_script)
         .arg("tests/remote-list-params.eval.ts")
@@ -366,7 +378,7 @@ fn eval_runner_list_mode_serializes_parameter_defaults() {
         .env("BT_EVAL_DEV_MODE", "list")
         .env("BT_EVAL_NO_SEND_LOGS", "1")
         .output()
-        .expect("run eval-runner.mts in list mode");
+        .expect("run eval-runner in list mode");
 
     if !output.status.success() {
         let stdout = String::from_utf8_lossy(&output.stdout);
@@ -698,11 +710,16 @@ fn collect_deno_eval_diagnostics(dir: &Path, files: &[String]) -> Option<String>
     }
 
     let root = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-    let runner_script = root.join("scripts").join("eval-runner.mts");
+    let impl_source = fs::read_to_string(root.join("scripts").join("eval-runner-impl.ts")).ok()?;
+    // Match what bt materializes for non-vite-node runners: impl body with
+    // fire-and-forget appended. Standalone, no relative import to resolve.
+    let standalone = format!(
+        "{impl_source}\nmain().catch((err) => {{\n  console.error(err);\n  process.exit(1);\n}});\n"
+    );
     let local_runner_dir = dir.join(".bt").join("eval-runners");
     fs::create_dir_all(&local_runner_dir).ok()?;
-    let local_runner = local_runner_dir.join("diag-eval-runner.mts");
-    fs::copy(&runner_script, &local_runner).ok()?;
+    let local_runner = local_runner_dir.join("diag-eval-runner.ts");
+    fs::write(&local_runner, &standalone).ok()?;
     let runner_script_str = local_runner.to_string_lossy().to_string();
 
     let mut cmd = Command::new("deno");
