+++ tsconfig.json
@@ -5,6 +5,7 @@
     "moduleResolution": "NodeNext",
     "strict": true,
     "noEmit": true,
+    "allowImportingTsExtensions": true,
     "types": ["node"],
     "resolveJsonModule": true,
     "esModuleInterop": true,
